+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -1,7 +1,9 @@
 const DatabaseTableName = {
 	ACTIVITIES: "activities",
+	ACTIVITY_LIKES: "activity_likes",
 	CHAT_MESSAGES: "chat_messages",
 	CHATS: "chats",
+	COMMENTS: "comments",
 	COURSE_SECTIONS: "course_sections",
 	COURSES: "courses",
 	COURSES_TO_USERS: "courses_to_users",