+++ apps/backend/src/modules/activities/activity.entity.ts
@@ -6,7 +6,11 @@ import { type ActivityType } from "./libs/enums/enums.js";
 class ActivityEntity implements Entity {
 	public actionId: number;
 
+	public commentCount: null | number;
+
 	public id: null | number;
+
+	public likesCount: null | number;
 
 	public payload: unknown;
 
@@ -20,7 +24,9 @@ class ActivityEntity implements Entity {
 
 	private constructor({
 		actionId,
+		commentCount,
 		id,
+		likesCount,
 		payload,
 		type,
 		updatedAt,
@@ -28,7 +34,9 @@ class ActivityEntity implements Entity {
 		userId,
 	}: {
 		actionId: number;
+		commentCount: null | number;
 		id: null | number;
+		likesCount: null | number;
 		payload: unknown;
 		type: ValueOf<typeof ActivityType>;
 		updatedAt: string;
@@ -37,16 +45,20 @@ class ActivityEntity implements Entity {
 	}) {
 		this.actionId = actionId;
 		this.id = id;
+		this.likesCount = likesCount;
 		this.payload = payload;
 		this.type = type;
 		this.updatedAt = updatedAt;
+		this.commentCount = commentCount;
 		this.user = user;
 		this.userId = userId;
 	}
 
 	public static initialize({
 		actionId,
+		commentCount,
 		id,
+		likesCount,
 		payload,
 		type,
 		updatedAt,
@@ -54,7 +66,9 @@ class ActivityEntity implements Entity {
 		userId,
 	}: {
 		actionId: number;
+		commentCount: null | number;
 		id: number;
+		likesCount: null | number;
 		payload: unknown;
 		type: ValueOf<typeof ActivityType>;
 		updatedAt: string;
@@ -63,7 +77,9 @@ class ActivityEntity implements Entity {
 	}): ActivityEntity {
 		return new ActivityEntity({
 			actionId,
+			commentCount,
 			id,
+			likesCount,
 			payload,
 			type,
 			updatedAt,
@@ -85,7 +101,9 @@ class ActivityEntity implements Entity {
 	}): ActivityEntity {
 		return new ActivityEntity({
 			actionId,
+			commentCount: null,
 			id: null,
+			likesCount: null,
 			payload,
 			type,
 			updatedAt: "",
@@ -129,6 +147,39 @@ class ActivityEntity implements Entity {
 		return {
 			actionId: this.actionId,
 			id: this.id as number,
+			payload: this.payload,
+			type: this.type,
+			updatedAt: this.updatedAt,
+			user: (this.user as UserEntity).toObject(),
+			userId: this.userId,
+		};
+	}
+
+	public toObjectWithRelationsAndCounts(): {
+		actionId: number;
+		commentCount: null | number;
+		id: number;
+		likesCount: null | number;
+		payload: unknown;
+		type: ValueOf<typeof ActivityType>;
+		updatedAt: string;
+		user: {
+			avatarUrl: string;
+			createdAt: string;
+			email: string;
+			firstName: string;
+			id: number;
+			lastName: string;
+			nickname: null | string;
+			updatedAt: string;
+		};
+		userId: number;
+	} {
+		return {
+			actionId: this.actionId,
+			commentCount: Number(this.commentCount),
+			id: this.id as number,
+			likesCount: Number(this.likesCount),
 			payload: this.payload,
 			type: this.type,
 			updatedAt: this.updatedAt,