+++ apps/backend/src/modules/comments/libs/enums/relation-name.enum.ts
@@ -1,0 +1,6 @@
+const RelationName = {
+	AUTHOR: "author",
+	AVATAR_FILE: "avatarFile",
+} as const;
+
+export { RelationName };