+++ packages/shared/src/modules/activities/libs/types/activity-counts.type.ts
@@ -1,0 +1,6 @@
+type ActivityCounts = {
+	commentCount: number;
+	likesCount: number;
+};
+
+export { type ActivityCounts };