+++ packages/shared/src/modules/comments/libs/types/comment-get-all-request-dto.type.ts
@@ -1,0 +1,5 @@
+type CommentGetAllRequestDto = {
+	activityId: number;
+};
+
+export { type CommentGetAllRequestDto };