+++ packages/shared/src/modules/comments/libs/types/types.ts
@@ -1,0 +1,5 @@
+export { type CommentCreateRequestDto } from "./comment-create-request-dto.type.js";
+export { type CommentGetAllRequestDto } from "./comment-get-all-request-dto.type.js";
+export { type CommentGetAllResponseDto } from "./comment-get-all-response-dto.type.js";
+export { type CommentUpdateRequestDto } from "./comment-update-request-dto.type.js";
+export { type CommentWithRelationsResponseDto } from "./comment-with-relations-response-dto.type.js";